misc: rename control to controllers and compiled_controller to compiled
