+++ tutorials/compiled_control/compiled_control.py
@@ -1,6 +1,6 @@
 import os
 
-from opensourceleg.control.compiled_controller import CompiledController
+from opensourceleg.controllers.compiled import CompiledController
 
 my_linalg = CompiledController(
     library_name="lin_alg.so",
